+++ src/FPath/AppendableFPath.hs
@@ -1,5 +1,5 @@
 module FPath.AppendableFPath
-  ( AppendableFPath( (⫻) ), (</>) )
+  ( AppendableFPath( AppendableFPathD, AppendableFPathF, (⫻) ), (</>) )
 where
 
 -- base-unicode-symbols ----------------
@@ -33,25 +33,35 @@ import FPath.RelFile  ( RelFile )
 
 infixl 7 ⫻
 
-class AppendableFPath α β γ | α β → γ where
-  (⫻) ∷ α → β → γ
+class AppendableFPath γ where
+  type AppendableFPathD γ
+  type AppendableFPathF γ
+  (⫻) ∷ AppendableFPathD γ → AppendableFPathF γ → γ
 
-instance AppendableFPath AbsDir RelDir AbsDir where
+instance AppendableFPath AbsDir where
+  type AppendableFPathD AbsDir = AbsDir
+  type AppendableFPathF AbsDir = RelDir
   d ⫻ f = (d ⊣ seq ⊕ f ⊣ seq) ⊣ re seq
 
-instance AppendableFPath AbsDir RelFile AbsFile where
+instance AppendableFPath AbsFile where
+  type AppendableFPathD AbsFile = AbsDir
+  type AppendableFPathF AbsFile = RelFile
   d ⫻ f = (d ⊣ seq ⪡ f ⊣ seqNE) ⊣ re seqNE
 
-instance AppendableFPath RelDir RelDir RelDir where
+instance AppendableFPath RelDir where
+  type AppendableFPathD RelDir = RelDir
+  type AppendableFPathF RelDir = RelDir
   d ⫻ f = (d ⊣ seq ⊕ f ⊣ seq) ⊣ re seq
 
-instance AppendableFPath RelDir RelFile RelFile where
+instance AppendableFPath RelFile where
+  type AppendableFPathD RelFile = RelDir
+  type AppendableFPathF RelFile = RelFile
   d ⫻ f = (d ⊣ seq ⪡ f ⊣ seqNE) ⊣ re seqNE
 
 infixl 7 </>
 
 -- | non-unicode synonym of `(⫻)`
-(</>) ∷ AppendableFPath α β γ ⇒ α → β → γ
+(</>) ∷ AppendableFPath γ ⇒ AppendableFPathD γ → AppendableFPathF γ → γ
 (</>) = (⫻)
 
 -- that's all, folks! ----------------------------------------------------------